refactor: use DrinkModel to display details screen
