+++ app/src/main/java/com/pereyrarg11/drinks/feature/drink/data/remote/DrinkDetailApi.kt
@@ -1,7 +1,7 @@
 package com.pereyrarg11.drinks.feature.drink.data.remote
 
+import com.pereyrarg11.drinks.core.data.remote.dto.DrinkDto
 import com.pereyrarg11.drinks.core.data.remote.dto.ResponseBodyDto
-import com.pereyrarg11.drinks.feature.drink.data.remote.dto.DrinkDetailDto
 import com.skydoves.sandwich.ApiResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
@@ -10,5 +10,5 @@ interface DrinkDetailApi {
     @GET("lookup.php")
     suspend fun getDrinkDetail(
         @Query("i") id: String,
-    ): ApiResponse<ResponseBodyDto<DrinkDetailDto>>
+    ): ApiResponse<ResponseBodyDto<DrinkDto>>
 }
