+++ app/src/main/java/com/pereyrarg11/drinks/feature/drink/data/repository/DrinkDetailRepositoryImpl.kt
@@ -1,12 +1,13 @@
 package com.pereyrarg11.drinks.feature.drink.data.repository
 
+import com.pereyrarg11.drinks.core.data.remote.dto.DrinkDto
 import com.pereyrarg11.drinks.core.data.remote.dto.ResponseBodyDto
 import com.pereyrarg11.drinks.core.data.util.Converter
 import com.pereyrarg11.drinks.core.data.util.NoResultsFoundException
+import com.pereyrarg11.drinks.core.di.EntityDetailConverter
+import com.pereyrarg11.drinks.core.domain.model.DrinkModel
 import com.pereyrarg11.drinks.core.domain.util.DataResult
 import com.pereyrarg11.drinks.feature.drink.data.remote.DrinkDetailApi
-import com.pereyrarg11.drinks.feature.drink.data.remote.dto.DrinkDetailDto
-import com.pereyrarg11.drinks.feature.drink.domain.model.DrinkDetailModel
 import com.pereyrarg11.drinks.feature.drink.domain.repository.DrinkDetailRepository
 import com.skydoves.sandwich.suspendOnError
 import com.skydoves.sandwich.suspendOnException
@@ -19,10 +20,10 @@ import javax.inject.Singleton
 @Singleton
 class DrinkDetailRepositoryImpl @Inject constructor(
     private val api: DrinkDetailApi,
-    private val responseConverter: Converter<ResponseBodyDto<DrinkDetailDto>, List<DrinkDetailModel>>,
+    @EntityDetailConverter private val responseConverter: Converter<ResponseBodyDto<DrinkDto>, List<DrinkModel>>,
 ) : DrinkDetailRepository {
 
-    override suspend fun getDrinkById(id: String): Flow<DataResult<DrinkDetailModel>> {
+    override suspend fun getDrinkById(id: String): Flow<DataResult<DrinkModel>> {
         return flow {
             emit(DataResult.Loading(isLoading = true))
