--- app/src/main/java/com/pereyrarg11/drinks/feature/drink/domain/model/DrinkMediaModel.kt
@@ -1,8 +1,0 @@
-package com.pereyrarg11.drinks.feature.drink.domain.model
-
-data class DrinkMediaModel(
-    val thumbUrl: String = "",
-    val imageUrl: String = "",
-    val videoUrl: String = "",
-    val attribution: String = "",
-)
