+++ app/src/main/java/com/pereyrarg11/drinks/feature/drink/domain/repository/DrinkDetailRepository.kt
@@ -1,9 +1,9 @@
 package com.pereyrarg11.drinks.feature.drink.domain.repository
 
+import com.pereyrarg11.drinks.core.domain.model.DrinkModel
 import com.pereyrarg11.drinks.core.domain.util.DataResult
-import com.pereyrarg11.drinks.feature.drink.domain.model.DrinkDetailModel
 import kotlinx.coroutines.flow.Flow
 
 interface DrinkDetailRepository {
-    suspend fun getDrinkById(id: String): Flow<DataResult<DrinkDetailModel>>
+    suspend fun getDrinkById(id: String): Flow<DataResult<DrinkModel>>
 }
