+++ app/src/main/java/com/pereyrarg11/drinks/feature/drink/domain/usecase/GetDrinkUseCase.kt
@@ -1,7 +1,7 @@
 package com.pereyrarg11.drinks.feature.drink.domain.usecase
 
+import com.pereyrarg11.drinks.core.domain.model.DrinkModel
 import com.pereyrarg11.drinks.core.domain.util.DataResult
-import com.pereyrarg11.drinks.feature.drink.domain.model.DrinkDetailModel
 import com.pereyrarg11.drinks.feature.drink.domain.repository.DrinkDetailRepository
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
@@ -10,6 +10,6 @@ class GetDrinkUseCase @Inject constructor(
     private val repository: DrinkDetailRepository,
 ) {
 
-    suspend operator fun invoke(id: String): Flow<DataResult<DrinkDetailModel>> =
+    suspend operator fun invoke(id: String): Flow<DataResult<DrinkModel>> =
         repository.getDrinkById(id)
 }
