+++ app/src/main/java/com/pereyrarg11/drinks/feature/drink/presentation/DrinkViewModel.kt
@@ -7,12 +7,13 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.viewModelScope
 import com.pereyrarg11.drinks.core.data.util.Converter
 import com.pereyrarg11.drinks.core.data.util.MissingParamsException
+import com.pereyrarg11.drinks.core.di.EntityDetailConverter
+import com.pereyrarg11.drinks.core.domain.model.DrinkModel
 import com.pereyrarg11.drinks.core.domain.util.DataResult
+import com.pereyrarg11.drinks.core.logger.error.ErrorLogger
 import com.pereyrarg11.drinks.core.presentation.BaseViewModel
 import com.pereyrarg11.drinks.core.presentation.navigation.NavConstants.ID_PARAM
-import com.pereyrarg11.drinks.core.logger.error.ErrorLogger
 import com.pereyrarg11.drinks.feature.drink.analytics.DrinkAnalyticsLogger
-import com.pereyrarg11.drinks.feature.drink.domain.model.DrinkDetailModel
 import com.pereyrarg11.drinks.feature.drink.domain.usecase.GetDrinkUseCase
 import com.pereyrarg11.drinks.feature.drink.presentation.model.DrinkUiDetail
 import com.pereyrarg11.drinks.feature.drink.presentation.model.DrinkUiState
@@ -26,7 +27,7 @@ class DrinkViewModel @Inject constructor(
     savedStateHandle: SavedStateHandle,
     errorLogger: ErrorLogger,
     val getDrink: GetDrinkUseCase,
-    val converter: Converter<DrinkDetailModel, DrinkUiDetail>,
+    @EntityDetailConverter val converter: Converter<DrinkModel, DrinkUiDetail>,
     private val analyticsLogger: DrinkAnalyticsLogger,
 ) : BaseViewModel(errorLogger) {
